+++ src/components/auth/LoginForm/login.form.module.css
@@ -1,0 +1,18 @@
+.form {
+  width: 100%;
+  max-width: 340px;
+  margin: 0 auto;
+  display: flex;
+  gap: 10px;
+  flex-direction: column;
+}
+
+.form_buttons {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.form_buttons button {
+  height: 56px;
+}
