+++ src/pages/Auth/Login/login.module.css
@@ -6,22 +6,3 @@
   margin: 0 auto;
   padding: 50px 0;
 }
-
-.form {
-  width: 100%;
-  max-width: 340px;
-  margin: 0 auto;
-  display: flex;
-  gap: 10px;
-  flex-direction: column;
-}
-
-.form_buttons {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.form_buttons button {
-  height: 56px;
-}
